// packages/wallet/core/src/relayer/standard/rpc/relayer.gen.ts
export enum FeeTokenType {
  UNKNOWN = 'UNKNOWN',
  ERC20_TOKEN = 'ERC20_TOKEN',
  ERC1155_TOKEN = 'ERC1155_TOKEN',
}
// ...
export interface IntentPrecondition {
  type: string
  chainId: string
  data: any
}
// ...
export interface FeeToken {
  chainId: number
  name: string
  symbol: string
  type: FeeTokenType
  decimals?: number
  logoURL: string
  contractAddress?: string
  tokenID?: string
}